+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -54,6 +54,9 @@ Notes](../../RELEASENOTES.md).
   thresholds are present.
   ([#7221](https://github.com/open-telemetry/opentelemetry-dotnet/issues/7221))
 
+* Fix `ArgumentException` if `OTEL_SDK_DISABLED=true`.
+  ([#TODO](https://github.com/open-telemetry/opentelemetry-dotnet/issues/TODO))
+
 ## 1.15.3-beta.1
 
 Released 2026-Apr-21